perplexity issue fixed
